Share causal LM/encoder/decoder forward method
This change introduces the `TransformerEncoder`, `TransformerDecoder`,
and `TransformerCausalLM` base classes that provide standard `forward`
methods. Use these base classes for all models that have a standard
`forward` method.

While at it, remove the padding identifier and implicit creation of an
attention mask. This behavior was quite confusing, since it was not used
by all models. Move to an explicit behavior: to use an attention mask,
specify it.

The RoBERTa model still uses padding_id, because it is needed for the
position embedding offset computation.
