+++ curated_transformers/models/bert/_hf.py
@@ -18,7 +18,6 @@ HF_KEY_TO_CURATED_KEY = MappingProxyType(
 
 
 HF_CONFIG_KEY_MAPPING = {
-    "pad_token_id": "padding_id",
     "attention_probs_dropout_prob": "attention_probs_dropout_prob",
     "hidden_act": "hidden_act",
     "hidden_dropout_prob": "hidden_dropout_prob",
