+++ curated_transformers/models/falcon/decoder.py
@@ -1,5 +1,5 @@
 from functools import partial
-from typing import Any, List, Mapping, Optional, Type, TypeVar
+from typing import Any, Mapping, Optional, Type, TypeVar
 
 import torch
 from torch import Tensor
@@ -8,11 +8,9 @@ from torch.nn import Dropout, Embedding, LayerNorm, ModuleList
 from ...layers.attention import (
     AttentionHeads,
     AttentionLinearBiases,
-    AttentionMask,
     QkvMode,
     SelfAttention,
 )
-from ...layers.cache import KeyValueCache
 from ...layers.embeddings import QueryKeyRotaryEmbeddings
 from ...layers.feedforward import PointwiseFeedForward
 from ...layers.transformer import (
@@ -21,8 +19,7 @@ from ...layers.transformer import (
     TransformerLayerNorms,
 )
 from ..hf_hub import FromHFHub
-from ..module import DecoderModule
-from ..output import ModelOutputWithCache
+from ..transformer import TransformerDecoder
 from ._hf import convert_hf_config, convert_hf_state_dict
 from .config import FalconConfig
 from .layer import OldFalconDecoderLayer
@@ -31,7 +28,7 @@ from .layer import OldFalconDecoderLayer
 Self = TypeVar("Self", bound="FalconDecoder")
 
 
-class FalconDecoder(DecoderModule, FromHFHub):
+class FalconDecoder(TransformerDecoder, FromHFHub):
     """
     `Falcon`_ decoder.
 
@@ -75,45 +72,6 @@ class FalconDecoder(DecoderModule, FromHFHub):
             config.layer.feedforward.hidden_width,
             config.layer.layer_norm_eps,
             device=device,
-        )
-
-    def forward(
-        self,
-        input_ids: Tensor,
-        attention_mask: Optional[AttentionMask] = None,
-        cache: Optional[List[KeyValueCache]] = None,
-        positions: Optional[Tensor] = None,
-        store_cache: bool = False,
-    ) -> ModelOutputWithCache[KeyValueCache]:
-        embeddings = self.embeddings(input_ids)
-        embeddings = self.dropout(embeddings)
-        layer_output = embeddings
-
-        layer_outputs = []
-        new_cache = []
-        layer_cache = None
-        for layer in self.layers:
-            if cache is not None:
-                # The key-value cache is stored per layer, so peel off one
-                # layer at a time.
-                layer_cache = cache[0]
-                cache = cache[1:]
-            layer_output, new_layer_cache = layer(
-                layer_output,
-                attention_mask,
-                cache=layer_cache,
-                store_cache=store_cache,
-                positions=positions,
-            )
-            layer_outputs.append(layer_output)
-            if store_cache:
-                new_cache.append(new_layer_cache)
-
-        layer_outputs[-1] = self.output_layer_norm(layer_outputs[-1])
-
-        return ModelOutputWithCache(
-            all_outputs=[embeddings, *layer_outputs],
-            cache=new_cache if store_cache else None,
         )
 
     @classmethod
